// frontend/src/lib/api/generated/models/ListingCommentRepliesItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ListingCommentRepliesItem {
    'id': number;
    'content': string;
    'createdAt': Date;
    'updatedAt': Date;
    'userId'?: number | null;
    'userName'?: string | null;
    'userAvatar'?: string | null;
    'userRole'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "userId",
            "baseName": "user_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "userName",
            "baseName": "user_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "userAvatar",
            "baseName": "user_avatar",
            "type": "string",
            "format": ""
        },
        {
            "name": "userRole",
            "baseName": "user_role",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingCommentRepliesItem.attributeTypeMap;
    }

    public constructor() {
    }
}